parser: enforce no newline before `<` in type references, and before `[` in lookup types and array types.
this is whitespace typescript cares about, and the non-whitespace aware parser in ST has had problems with it.
now I discovered a hack with using `Positional` to inspect the preceding line and make a decision based on that.

As a bonus, the very hacky concept of parser patches is now gone