DefaultedTypeArguments: guard against loop
for every type reference we look up what it points to, and then check if there are defaulted type arguments.
those type arguments in turn, in this particular case, happened to point to itself. Again it was missing defaulted type arguments, and so it went on.

this could be solved more precisely, but for now let's chuck an `any` in there.

Fixes #483